Split inner classes into separated files
